Production Version of BHAT
This is a working production version with a very simple pseudo-database. Mr Hai really encouraged me to implement a full Flask DB for this, and actually wrote a bunch of code for me to leverage, but his code is yet beyond me, even just to implement and troubleshoot. I will continue to learn about DBs and Flask, but Day 40 on this "quick' sidequest to practice Python is adding up for me and it is time to return to the main questline in preparation to write the PNPT this Christmas. Hope you all enjoy the story!